ztf_api/pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
name = "alerts-api"
version = "1.3.0"
description = "AlertsAPI"
authors = ["ALeRCE"]
readme = "README.md"
packages = [{include = "api", from = "src"}, {include = "core", from = "src"}, {include = "shared", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
flask-restx = "^1.1.0"
prometheus-flask-exporter = "^0.22.4"
returns = "^0.19.0"
dependency-injector = {version = "^4.41.0", extras = ["yaml"]}
db-plugins = { git = "https://git@github.com/alercebroker/db-plugins", branch = "main" }
ralidator_flask = { git = "https://git@github.com/alercebroker/ralidator-flask", rev = "50825f9"}
flask-cors = "^4.0.0"
psycopg2-binary = "^2.9.6"
fastapi = "0.100.0"
uvicorn = "0.23.1"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-docker = "^1.0.1"
pytest-cov = "^4.1.0"
behave = "^1.2.6"
requests = "^2.31.0"
pyjwt = "2.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
memray = "^1.8.1"


[tool.poetry.group.production.dependencies]
gunicorn = { version = "^21.0.1" }
psycogreen = "^1.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
